backend/pyproject.toml
[tool.poetry]
name = "streamlit_fastapi_togetherai_llama"
version = "0.0.1"
description = "A simple example showcasing streamlit, fastapi, togetherai and llamaindex"
authors = [
    "Sergiu Soima, Liana Soima, Stefania Mocan",
]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.9.2"
restack-ai = "^0.0.28"
fastapi = "^0.115.4"
llama-index = "^0.11.22"
llama-index-llms-together = "^0.2.0"
uvicorn = "^0.32.0"
python-dotenv = "0.19"
streamlit = "^1.40.0"
requests = "^2.32.3"
chromadb = "^0.5.18"
llama-stack = "^0.0.52"
llama-stack-client = "^0.0.50"
together = "^1.3.3"
openpyxl = "^3.1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
services = "src.services:run_services"
app = "src.app:run_app"
